Move parkingZoneMap under addressSelector
Refs PV-196
