+++ src/common/addressSelector/parkingZoneMap/parkingZonesMap.css
@@ -1,5 +1,4 @@
 .leaflet-container {
     height: 30em;
     width: 100%;
-    /*z-index: -1;*/
 }
